Convert dialog fragment to fragment and add placeholder layout
