+++ ground/src/main/res/navigation/nav_graph.xml
@@ -236,10 +236,10 @@
       app:popUpTo="@id/navGraph"
       app:popUpToInclusive="true" />
   </fragment>
-  <dialog
+  <fragment
     android:id="@+id/surveySelectorDialogFragment"
-    android:name="com.google.android.ground.ui.surveyselector.SurveySelectorDialogFragment"
-    android:label="ProjectSelectorDialogFragment" />
+    android:name="com.google.android.ground.ui.surveyselector.SurveySelectorFragment"
+    android:label="SurveySelectorFragment" />
   <dialog
     android:id="@+id/locationOfInterestSelectorFragment"
     android:name="com.google.android.ground.ui.home.locationofinterestselector.LocationOfInterestSelectorFragment"
