+++ ground/src/test/java/com/google/android/ground/ui/mapcontainer/PolygonDrawingInfoDialogFragmentTest.kt
@@ -22,7 +22,7 @@ import com.google.android.ground.BaseHiltTest
 import com.google.android.ground.MainActivity
 import com.google.android.ground.R
 import com.google.android.ground.ui.home.mapcontainer.PolygonDrawingInfoDialogFragment
-import com.google.android.ground.ui.surveyselector.SurveySelectorDialogFragment
+import com.google.android.ground.ui.surveyselector.SurveySelectorFragment
 import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Before
@@ -55,7 +55,7 @@ class PolygonDrawingInfoDialogFragmentTest : BaseHiltTest() {
     polygonDrawingInfoDialogFragment = PolygonDrawingInfoDialogFragment(onClickRunnable)
     polygonDrawingInfoDialogFragment.showNow(
       activity.supportFragmentManager,
-      SurveySelectorDialogFragment::class.java.simpleName
+      SurveySelectorFragment::class.java.simpleName
     )
     shadowOf(Looper.getMainLooper()).idle()
   }
